[InstCombine][DebugInfo] Remove redundant dbg-instrs during InstCombine
For certain inputs, InstCombine is prone to producing an incredibly
large number of debug intrinsics that are redundant, due to them all
being undef and/or appearing in a contiguous block with no instructions
in-between. This patch attempts to resolve this by removing redundant
debug instructions after each iteration of InstCombine, preventing the
explosion of intrinsics and massively reducing the size of the resulting
IR, improving performance for subsequent stages.