[CodeGen] Correctly handle non-standard cases in RemoveLoadsIntoFakeUses
In the RemoveLoadsIntoFakeUses pass, we try to remove loads that are only
used by fake uses, as well as the fake use in question. There are two
existing errors with the pass however: it incorrectly examines every operand
of each FAKE_USE, when only the first is relevant (extra operands should
just be "killed" regs), and it ignores cases where the FAKE_USE register is
not an exact match for the loaded register, which is incorrect as regalloc
may choose to load a wider value than was required pre-regalloc.

This patch fixes both of these cases.