[DebugInfo][InstCombine] Keep valid source locations in InstCombineNegator
Currently in InstCombineNegator, at the very end we call
IRBuilder::Insert on all the new instructions, with an empty DebugLoc()
set. The InsertPoint is empty at this point, so the purpose of this is to
invoke the insertion callback code, but stripping the DebugLoc from the
IRBuilder is unnecessary - the existing DebugLoc in the builder should
correspond to the root instruction being replaced, and so should be used
iff the negator has not assigned a more specific DebugLoc to any of the new
instructions.

Found using https://github.com/llvm/llvm-project/pull/107279.